#16 Moves tags to a separate trait
